made some cosmetic changes

+++ src/IO/MyCompressorOutputStream.java
@@ -19,7 +19,7 @@ public class MyCompressorOutputStream extends OutputStream {
     @Override
     public void write(int b) throws IOException {
 
-        out.write(((Integer) b).byteValue());
+        out.write((byte)b);
     }
 
     /**
